fix: reduce utils methods
