+++ mirai-core-mock/src/internal/contact/MockFriendImpl.kt
@@ -28,7 +28,6 @@ import net.mamoe.mirai.mock.internal.msgsrc.OnlineMsgSrcFromFriend
 import net.mamoe.mirai.mock.internal.msgsrc.OnlineMsgSrcToFriend
 import net.mamoe.mirai.mock.internal.msgsrc.newMsgSrc
 import net.mamoe.mirai.mock.utils.broadcastBlocking
-import net.mamoe.mirai.mock.utils.randomMockImage
 import net.mamoe.mirai.utils.ExternalResource
 import net.mamoe.mirai.utils.cast
 import net.mamoe.mirai.utils.lateinitMutableProperty
@@ -50,7 +49,7 @@ internal class MockFriendImpl(
 
         override var nick: String = nick
         override var remark: String = remark
-        override var avatarUrl: String by lateinitMutableProperty { randomMockImage(bot).getUrl(bot) }
+        override var avatarUrl: String by lateinitMutableProperty { MockImage.random(bot).getUrl(bot) }
     }
     override var avatarUrl: String
         get() = mockApi.avatarUrl
