+++ mirai-core-mock/src/internal/contact/MockStrangerImpl.kt
@@ -22,7 +22,6 @@ import net.mamoe.mirai.mock.contact.MockStranger
 import net.mamoe.mirai.mock.internal.msgsrc.OnlineMsgSrcFromStranger
 import net.mamoe.mirai.mock.internal.msgsrc.OnlineMsgSrcToStranger
 import net.mamoe.mirai.mock.internal.msgsrc.newMsgSrc
-import net.mamoe.mirai.mock.utils.randomMockImage
 import net.mamoe.mirai.utils.cast
 import net.mamoe.mirai.utils.lateinitMutableProperty
 import java.util.concurrent.CancellationException
@@ -42,7 +41,7 @@ internal class MockStrangerImpl(
         override var nick: String = nick
         override var remark: String = remark
     }
-    override var avatarUrl: String by lateinitMutableProperty { randomMockImage(bot).getUrl(bot) }
+    override var avatarUrl: String by lateinitMutableProperty { MockImage.random(bot).getUrl(bot) }
 
     override val nick: String
         get() = mockApi.nick
