+++ mirai-core-mock/src/internal/contact/util.kt
@@ -11,6 +11,7 @@
 
 package net.mamoe.mirai.mock.internal.contact
 
+import kotlinx.coroutines.runBlocking
 import net.mamoe.mirai.Bot
 import net.mamoe.mirai.contact.Group
 import net.mamoe.mirai.contact.Member
@@ -22,9 +23,10 @@ import net.mamoe.mirai.message.data.*
 import net.mamoe.mirai.mock.MockBot
 import net.mamoe.mirai.mock.contact.MockGroup
 import net.mamoe.mirai.mock.utils.mock
-import net.mamoe.mirai.utils.ExternalResource
-import net.mamoe.mirai.utils.plusHttpSubpath
-import net.mamoe.mirai.utils.toUHexString
+import net.mamoe.mirai.mock.utils.randomImageContent
+import net.mamoe.mirai.utils.*
+import java.util.*
+import kotlin.io.path.outputStream
 
 internal fun Member.requireBotPermissionHigherThanThis(msg: String) {
     if (this.permission < this.group.botPermission) return
@@ -98,6 +100,23 @@ internal class MockImage(
     override val size: Long = 0,
     override val imageType: ImageType = ImageType.UNKNOWN,
 ) : DeferredOriginUrlAware, Image {
+
+    companion object {
+        // create a mockImage with random content
+        internal fun random(bot: MockBot): MockImage {
+            val text = randomImageContent()
+            val bindId = "image/" + generateUUID(text.md5())
+            val uuid = "${System.currentTimeMillis()}-${UUID.randomUUID()}"
+            bot.tmpFsServer.fsSystem.getPath(uuid).outputStream().use { fso ->
+                fso.write(text)
+            }
+            runBlocking {
+                bot.tmpFsServer.bindFile(uuid, bindId)
+            }
+            return MockImage(generateImageId(text.md5()), bindId)
+        }
+    }
+
     private val _stringValue: String? by lazy(LazyThreadSafetyMode.NONE) { "[mirai:image:$imageId]" }
 
     override fun getUrl(bot: Bot): String {
